Move enum tests to tests/types/test_enum.py
